+++ gkraken/device/device_settings.py
@@ -1,17 +1,17 @@
 #  This file is part of gkraken.
-# 
+#
 #  Copyright (c) 2021 Roberto Leinardi and Guy Boldon
-# 
+#
 #  gkraken is free software: you can redistribute it and/or modify
 #  it under the terms of the GNU General Public License as published by
 #  the Free Software Foundation, either version 3 of the License, or
 #  (at your option) any later version.
-# 
+#
 #  gkraken is distributed in the hope that it will be useful,
 #  but WITHOUT ANY WARRANTY; without even the implied warranty of
 #  MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
 #  GNU General Public License for more details.
-# 
+#
 #  You should have received a copy of the GNU General Public License
 #  along with gkraken.  If not, see <http://www.gnu.org/licenses/>.
 
@@ -63,7 +63,7 @@ class DeviceSettings:
     _modes_ring: List[LightingMode] = []
 
     @classmethod
-    def determine_status(cls, status_list: list) -> Optional[Status]:
+    def determine_status(cls, status_list: list, device_description: str) -> Optional[Status]:
         """creates a Status object from the given liquidctl status_list"""
         raise NotImplementedError('This should be implemented in one of the child classes')