+++ gkraken/interactor/udev_interactor.py
@@ -22,18 +22,25 @@ from injector import singleton, inject
 from gkraken.interactor import _run_and_get_stdout
 
 _LOG = logging.getLogger(__name__)
-_UDEV_RULE = '''# Kraken X2 products
+_UDEV_RULE = '''# UDev rules that grant user access to GKraken for supported devices
+
+# Asetek 690LC (assuming NZXT Kraken X)
+# Asetek 690LC (assuming EVGA CLC)
+SUBSYSTEMS=="usb", ATTRS{idVendor}=="2433", ATTRS{idProduct}=="b200", MODE="0666"
+
+# NZXT Kraken X (X42, X52, X62 or X72)
 SUBSYSTEMS=="usb", ATTRS{idVendor}=="1e71", ATTRS{idProduct}=="170e", MODE="0666"
 
-# Kraken X3 products
+# NZXT Kraken X (X53, X63 or X73)
 SUBSYSTEMS=="usb", ATTRS{idVendor}=="1e71", ATTRS{idProduct}=="2007", MODE="0666"
 
-# Kraken Z3 products
+# NZXT Kraken Z (Z53, Z63 or Z73)
 SUBSYSTEMS=="usb", ATTRS{idVendor}=="1e71", ATTRS{idProduct}=="3008", MODE="0666"
 '''
 _UDEV_RULE_FILE_PATH = '/lib/udev/rules.d/60-gkraken.rules'
 _UDEV_RULE_RELOAD_COMMANDS = 'udevadm control --reload-rules ' \
-                             '&& udevadm trigger --subsystem-match=usb --attr-match=idVendor=1e71 --action=add'
+                             '&& udevadm trigger --subsystem-match=usb --attr-match=idVendor=1e71 --action=add' \
+                             '&& udevadm trigger --subsystem-match=usb --attr-match=idVendor=2433 --action=add'
 
 
 @singleton
@@ -50,10 +57,10 @@ class UdevInteractor:
                f'echo \'{_UDEV_RULE}\' > {_UDEV_RULE_FILE_PATH} && {_UDEV_RULE_RELOAD_COMMANDS}']
         result = _run_and_get_stdout(cmd)
         if result[0] != 0:
-            _LOG.warning(f"Error while creating rule file. Exit code: {result[0]}. {result[1]}")
+            _LOG.warning("Error while creating rule file. Exit code: %d. %s", result[0], result[1])
         else:
             _LOG.info("udev rules successfully applied. You may need to restart for changes to take effect.")
-        return result[0]
+        return int(result[0])
 
     @staticmethod
     def remove_udev_rule() -> int:
@@ -63,5 +70,7 @@ class UdevInteractor:
                f'rm {_UDEV_RULE_FILE_PATH} && {_UDEV_RULE_RELOAD_COMMANDS}']
         result = _run_and_get_stdout(cmd)
         if result[0] != 0:
-            _LOG.warning(f"Error while removing rule file. Exit code: {result[0]}. {result[1]}")
-        return result[0]
+            _LOG.warning("Error while removing rule file. Exit code: %d. %s", result[0], result[1])
+        else:
+            _LOG.info("udev rules successfully removed.")
+        return int(result[0])